1.99.21-beta / filesystem bugfixes

+++ includes/class-filesystem-translationfile.php
@@ -11,46 +11,7 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 	function __construct($type = 'unknown') {
 		parent::__construct($type);
 	}
-	
-	function create_pofile($pofile, $base_file, $proj_id, $timestamp, $translator, $pluralforms, $language, $country) {
-		global $wp_filesystem, $parent_file;
-		$current_parent  = $parent_file;
-		$parent_file 	 = 'tools.php'; //needed for screen icon :-)
-		if (function_exists('set_current_screen')) set_current_screen('tools'); //WP 3.0 fix
-					
-		//check the file system
-		ob_start();
-		$url = 'admin-ajax.php';
-		if ( false === ($credentials = request_filesystem_credentials($url)) ) {
-			$data = ob_get_contents();
-			ob_end_clean();
-			if( ! empty($data) ){
-				header('Status: 401 Unauthorized');
-				header('HTTP/1.1 401 Unauthorized');
-				echo $data;
-				exit;
-			}
-			return;
-		}
 
-		if ( ! WP_Filesystem($credentials) ) {
-			request_filesystem_credentials($url, '', true); //Failed to connect, Error and request again
-			$data = ob_get_contents();
-			ob_end_clean();
-			if( ! empty($data) ){
-				header('Status: 401 Unauthorized');
-				header('HTTP/1.1 401 Unauthorized');
-				echo $data;
-				exit;
-			}
-			return;
-		}
-		ob_end_clean();
-		$parent_file = $current_parent;
-		
-		return parent::create_pofile($pofile, $base_file, $proj_id, $timestamp, $translator, $pluralforms, $language, $country);
-	}
-	
 	function destroy_pofile($pofile) {
 		global $wp_filesystem, $parent_file;
 		$current_parent  = $parent_file;
@@ -86,9 +47,15 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 		}
 		ob_end_clean();
 		$parent_file = $current_parent;	
-	
+		
 		$error = false;
-		if (file_exists($pofile)) if (!@unlink($pofile)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $pofile);
+		if($wp_filesystem->method == 'direct') {
+			if (file_exists($pofile)) if (!@unlink($pofile)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $mofile);
+		}else {
+			$root_dir = str_replace('\\', '/', $_SERVER['DOCUMENT_ROOT']);
+			$target_file = str_replace($root_dir, '', $pofile);			
+			if($wp_filesystem->is_file($target_file)) if (!$wp_filesystem->delete($target_file)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $pofile);
+		}
 		if ($error) {
 			header('Status: 404 Not Found');
 			header('HTTP/1.1 404 Not Found');
@@ -134,7 +101,13 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 		$parent_file = $current_parent;
 	
 		$error = false;
-		if (file_exists($mofile)) if (!@unlink($mofile)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $mofile);
+		if($wp_filesystem->method == 'direct') {
+			if (file_exists($mofile)) if (!@unlink($mofile)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $mofile);
+		}else {
+			$root_dir = str_replace('\\', '/', $_SERVER['DOCUMENT_ROOT']);
+			$target_file = str_replace($root_dir, '', $mofile);			
+			if($wp_filesystem->is_file($target_file)) if (!$wp_filesystem->delete($target_file)) $error = sprintf(__("You do not have the permission to delete the file '%s'.", CSP_PO_TEXTDOMAIN), $mofile);
+		}
 		if ($error) {
 			header('Status: 404 Not Found');
 			header('HTTP/1.1 404 Not Found');
@@ -178,19 +151,37 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 		}
 		ob_end_clean();
 		$parent_file = $current_parent;
-		
+	
 		$error = false;
-		if (file_exists($filename)) {
-			@chmod($filename, 0644);
-			if(!is_writable($filename)) {
-				@chmod($filename, 0664);
-				if (!is_writable($filename)) {
-					@chmod($filename, 0666);
+		if($wp_filesystem->method == 'direct') {		
+			if (file_exists($filename)) {
+				@chmod($filename, 0644);
+				if(!is_writable($filename)) {
+					@chmod($filename, 0664);
+					if (!is_writable($filename)) {
+						@chmod($filename, 0666);
+					}
+					if (!is_writable($filename)) $error = __('Server Restrictions: Changing file rights is not permitted.', CSP_PO_TEXTDOMAIN);
 				}
-				if (!is_writable($filename)) $error = __('Server Restrictions: Changing file rights is not permitted.', CSP_PO_TEXTDOMAIN);
 			}
+			else $error = sprintf(__("You do not have the permission to modify the file rights for a not existing file '%s'.", CSP_PO_TEXTDOMAIN), $filename);
+		} else {
+			$root_dir = str_replace('\\', '/', $_SERVER['DOCUMENT_ROOT']);
+			$target_file = str_replace($root_dir, '', $filename);			
+			if($wp_filesystem->is_file($target_file)) {
+				$wp_filesystem->chmod($target_file, 0644);
+				if(!is_writable($filename)) {
+					$wp_filesystem->chmod($target_file, 0664);
+					if (!is_writable($filename)) {
+						$wp_filesystem->chmod($target_file, 0666);
+					}
+					if (!is_writable($filename)) $error = __('Server Restrictions: Changing file rights is not permitted.', CSP_PO_TEXTDOMAIN);
+				}
+			}else{
+				$error = sprintf(__("You do not have the permission to modify the file rights for a not existing file '%s'.", CSP_PO_TEXTDOMAIN), $filename);
+			}
+			
 		}
-		else $error = sprintf(__("You do not have the permission to modify the file rights for a not existing file '%s'.", CSP_PO_TEXTDOMAIN), $filename);
 		if ($error) {
 			header('Status: 404 Not Found');
 			header('HTTP/1.1 404 Not Found');
@@ -234,8 +225,15 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 		}
 		ob_end_clean();
 		$parent_file = $current_parent;
+
+		if($wp_filesystem->method == 'direct') {
+			return parent::write_pofile($pofile, $last, $textdomain, $tds);
+		}else{
+			$root_dir = str_replace('\\', '/', $_SERVER['DOCUMENT_ROOT']);
+			$target_file = str_replace($root_dir, '', $pofile);			
+			return $wp_filesystem->put_contents($target_file, parent::ftp_get_pofile_content($pofile, $last, $textdomain, $tds));
+		}
 		
-		return parent::write_pofile($pofile, $last, $textdomain, $tds);
 	}
 
 	function write_mofile($mofile, $textdomain) {
@@ -274,7 +272,14 @@ class CspFileSystem_TranslationFile extends CspTranslationFile {
 		ob_end_clean();
 		$parent_file = $current_parent;
 		
-		return parent::write_mofile($mofile, $textdomain);
+		if($wp_filesystem->method == 'direct') {
+			return parent::write_mofile($mofile, $textdomain);
+		}else{
+			$root_dir = str_replace('\\', '/', $_SERVER['DOCUMENT_ROOT']);
+			$target_file = str_replace($root_dir, '', $mofile);
+			return $wp_filesystem->put_contents($target_file, parent::ftp_get_mofile_content($mofile, $textdomain));
+		}
+		
 	}
 	
 }
